backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.20.0
pydantic==2.10.3
pydantic-settings==2.7.0
python-multipart==0.0.18
pyyaml==6.0.2
jinja2==3.1.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.28.1
